Move constants to core
